+++ lib/nori/core_ext/object.rb
@@ -2,6 +2,12 @@ module Nori
   module CoreExt
     module Object
 
+      def xml_attributes
+        @xml_attributes ||= {}
+      end
+
+      attr_writer :xml_attributes
+
       def blank?
         respond_to?(:empty?) ? empty? : !self
       end unless method_defined?(:blank?)