+++ lib/nori/string_io_file.rb
@@ -1,7 +1,0 @@
-module Nori
-  class StringIOFile < StringIO
-
-    attr_accessor :original_filename, :content_type
-
-  end
-end