+++ lib/nori/string_with_attributes.rb
@@ -1,7 +1,0 @@
-module Nori
-  class StringWithAttributes < String
-
-    attr_accessor :attributes
-
-  end
-end