+++ lib/nori/xml_utility_node.rb
@@ -4,9 +4,6 @@ require "time"
 require "yaml"
 require "bigdecimal"
 
-require "nori/string_with_attributes"
-require "nori/string_io_file"
-
 module Nori
 
   # This is a slighly modified version of the XMLUtilityNode from
@@ -87,14 +84,14 @@ module Nori
 
     attr_accessor :name, :attributes, :children, :type
 
-    def prefixed_attributes
+    def prepared_attributes
       attributes.inject({}) do |memo, (key, value)|
-        memo[prefixed_attribute_name("@#{key}")] = value
+        memo[prepared_attribute_name(key)] = value
         memo
       end
     end
 
-    def prefixed_attribute_name(attribute)
+    def prepared_attribute_name(attribute)
       @nori.convert_tags? ? @nori.convert_tag(attribute) : attribute
     end
 
@@ -105,20 +102,18 @@ module Nori
 
     def to_hash
       if @type == "file"
-        f = StringIOFile.new((@children.first || '').unpack('m').first)
-        f.original_filename = attributes['name'] || 'untitled'
-        f.content_type = attributes['content_type'] || 'application/octet-stream'
+        f = StringIO.new((@children.first || '').unpack('m').first)
+        f.xml_attributes = {
+          :original_filename => (attributes['name'] || 'untitled'),
+          :content_type => (attributes['content_type'] || 'application/octet-stream')
+        }
         return { name => f }
       end
 
       if @text
         t = typecast_value unnormalize_xml_entities(inner_html)
         t = advanced_typecasting(t) if t.is_a?(String) && @nori.advanced_typecasting?
-
-        if t.is_a?(String)
-          t = StringWithAttributes.new(t)
-          t.attributes = attributes
-        end
+        t.xml_attributes = attributes
 
         return { name => t }
       else
@@ -146,15 +141,13 @@ module Nori
               out.merge!( k => v.map{|e| e.to_hash[k]})
             end
           end
-          out.merge! prefixed_attributes unless attributes.empty?
           out = out.empty? ? nil : out
         end
 
-        if @type && out.nil?
-          { name => typecast_value(out) }
-        else
-          { name => out }
-        end
+        out = typecast_value(out) if @type && out.nil?
+        out.xml_attributes = prepared_attributes unless attributes.empty?
+
+        { name => out }
       end
     end